# docs/swagger.yaml
basePath: /
definitions:
...
  app.ListRes:
    properties:
      list:
        description: Data list // 数据清单
      pager:
        allOf:
        - $ref: '#/definitions/app.Pager'
        description: Pagination info // 翻页信息
    type: object
  app.Pager:
    properties:
      page:
        description: Page number // 页码
        type: integer
      pageSize:
        description: Page size // 每页数量
        type: integer
      totalRows:
        description: Total rows // 总行数
        type: integer
    type: object
  app.Res:
    properties:
      code:
        type: integer
      context: {}
      data: {}
      details: {}
      message: {}
      pageIndex:
        description: '1-BASED wire value: 0/absent = non-paginated message, n>0 =
          download page n-1 (internal 0-based). 1-based so page 0 stays distinguishable
          from non-paginated messages under proto3 zero-value elision; only set for
          paginated download messages on pv>=2 connections. Client-to-server PageAck.pageIndex
          stays 0-based. // 线上值为 1-BASED：0/缺省 = 非分页消息，n>0 = 下载页 n-1（内部 0-based）。采用
          1-based 是为了让第 0 页在 proto3 零值不编码规则下仍可与非分页消息区分；仅 pv>=2 连接的分页下载消息会设置。客户端→服务端的
          PageAck.pageIndex 保持 0-based'
      path: {}
      status:
        type: boolean
      vault: {}
    type: object
...
  dto.NoteDTO:
    properties:
      clientName:
        description: Client name // 客户端名称
        type: string
      clientType:
        description: Client type // 客户端类型
        type: string
      clientVersion:
        description: Client version // 客户端版本
        type: string
      content:
        description: Note content // 笔记内容
        type: string
      contentHash:
        description: Content hash // 内容哈希
        type: string
      createdAt:
        description: Created at time // 创建时间
        type: string
      ctime:
        description: Creation timestamp // 创建时间戳
        type: integer
      entryId:
        description: v3 entry UUID (fs_entry.id) // v3 条目 UUID
        type: string
      id:
        description: Note ID // 笔记 ID
        type: integer
      lastTime:
        description: Record update timestamp // 记录更新时间戳
        type: integer
      mtime:
        description: Modification timestamp // 修改时间戳
        type: integer
      path:
        description: Note path // 笔记路径
        type: string
      pathHash:
        description: Path hash // 路径哈希
        type: string
      size:
        description: Note size // 笔记大小
        type: integer
      updatedAt:
        description: Updated at time // 更新时间
        type: string
      version:
        description: Version number // 版本号
        type: integer
    type: object
  dto.NoteHistoryDTO:
    properties:
      clientName:
        description: Client that made changes // 产生变更的客户端
        type: string
      clientType:
        description: Client type // 客户端类型
        type: string
      clientVersion:
        description: Client version // 客户端版本
        type: string
      content:
        description: Full historical content // 完整历史内容
        type: string
      contentHash:
        description: Content hash // 内容哈希
        type: string
      createdAt:
        description: Creation time of this version // 此版本的创建时间
        type: string
      diffs:
        description: Text differences // 文本差异内容
        items:
          type: object
        type: array
      entryId:
        description: v3 entry UUID (fs_entry.id) // v3 条目 UUID
        type: string
      id:
        description: History entry ID // 历史项 ID
        type: integer
      noteId:
        description: Associated note ID // 笔记 ID
        type: integer
      path:
        description: Note path at that time // 当时的笔记路径
        type: string
      vaultId:
        description: Associated vault ID // 保险库 ID
        type: integer
      version:
        description: Historical version number // 历史版本号
        type: integer
    type: object
  dto.NoteHistoryRestoreRequest:
    properties:
      historyId:
        description: History version ID // 历史版本 ID
        example: 1
        type: integer
      vault:
        description: Vault name // 保险库名称
        example: MyVault
        type: string
    required:
    - historyId
    - vault
    type: object
...
  dto.SettingDTO:
    properties:
      content:
        description: Setting content // 配置内容
        type: string
      contentHash:
        description: Content hash // 内容哈希
        type: string
      createdAt:
        description: Created at time // 创建时间
        type: string
      ctime:
        description: Creation timestamp // 创建时间戳
        type: integer
      id:
        description: Setting ID // 配置 ID
        type: integer
      lastTime:
        description: Record update timestamp // 记录更新时间戳
        type: integer
      mtime:
        description: Modification timestamp // 修改时间戳
        type: integer
      path:
        description: Setting path // 配置路径
        type: string
      pathHash:
        description: Path hash // 路径哈希值
        type: string
      updatedAt:
        description: Updated at time // 更新时间
        type: string
    type: object
  dto.SettingDeleteRequest:
    properties:
      context:
        description: Context // 同步上下文
        example: ctx123
        type: string
      path:
        description: Setting path // 配置路径
        example: User/Theme
        type: string
      pathHash:
        description: Path hash // 路径哈希
        example: hash123
        type: string
      vault:
        description: Vault name // 保险库名称
        example: MyVault
        type: string
    required:
    - path
    - vault
    type: object
  dto.SettingModifyOrCreateRequest:
    properties:
      content:
        description: Setting content // 配置内容
        example: dark
        type: string
      contentHash:
        description: Content hash // 内容哈希
        example: chash456
        type: string
      context:
        description: Context // 同步上下文
        example: ctx123
        type: string
      ctime:
        description: Creation timestamp // 创建时间戳
        example: 1700000000
        type: integer
      mtime:
        description: Modification timestamp // 修改时间戳
        example: 1700000000
        type: integer
      path:
        description: Setting path // 配置路径
        example: User/Theme
        type: string
      pathHash:
        description: Path hash // 路径哈希
        example: hash123
        type: string
      vault:
        description: Vault name // 保险库名称
        example: MyVault
        type: string
    required:
    - path
    - vault
    type: object
  dto.SettingRenameRequest:
    properties:
      newPath:
        description: New path // 新路径
        example: New/Path
        type: string
      newPathHash:
        description: New path hash // 新路径哈希
        example: newhash456
        type: string
      oldPath:
        description: Old path // 旧路径
        example: Old/Path
        type: string
      oldPathHash:
        description: Old path hash // 旧路径哈希
        example: oldhash123
        type: string
      vault:
        description: Vault name // 保险库名称
        example: MyVault
        type: string
    required:
    - newPath
    - oldPath
    - vault
    type: object
...
host: localhost:9000
info:
  contact:
    email: haierkeys@gmail.com
    name: Haierkeys
    url: https://github.com/haierkeys
  description: This is the Fast Note Sync Service HTTP API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Fast Note Sync Service HTTP API
  version: "1.0"
paths:
...
  /api/note/histories:
    get:
      description: Get all history records for a specific note with pagination
      parameters:
      - description: Is in recycle bin // 是否在回收站
        example: false
        in: query
        name: isRecycle
        type: boolean
      - description: Note path // 笔记路径
        example: ReadMe.md
        in: query
        name: path
        required: true
        type: string
      - description: Path hash // 路径哈希
        example: hash123
        in: query
        name: pathHash
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      - description: Page number // 页码
        in: query
        name: page
        type: integer
      - description: Page size // 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/app.ListRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.NoteHistoryDTO'
                        type: array
                    type: object
              type: object
      security:
      - UserAuthToken: []
      summary: Get note history list
      tags:
      - Note History
  /api/note/history:
    get:
      description: Get specific note history content by history record ID
      parameters:
      - description: History Record ID
        format: int64
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoteHistoryDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Get note history details
      tags:
      - Note History
  /api/note/history/restore:
    put:
      consumes:
      - application/json
      description: Restore note content to a specific history version
      parameters:
      - description: Restore Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.NoteHistoryRestoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoteDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Restore note from history
      tags:
      - Note History
...
  /api/setting:
    delete:
      consumes:
      - application/json
      description: Soft delete a setting by path or pathHash
      parameters:
      - description: Delete Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.SettingDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Delete setting
      tags:
      - Setting
    get:
      description: Get setting info for current user by path or pathHash
      parameters:
      - description: Setting path // 配置路径
        example: User/Theme
        in: query
        name: path
        type: string
      - description: Path hash // 路径哈希
        example: hash123
        in: query
        name: pathHash
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.SettingDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Get setting info
      tags:
      - Setting
    post:
      consumes:
      - application/json
      description: Create a new setting or update an existing one
      parameters:
      - description: Create/Update Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.SettingModifyOrCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.SettingDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Create or update setting
      tags:
      - Setting
  /api/setting/rename:
    post:
      consumes:
      - application/json
      description: Rename a setting and update its path and pathHash
      parameters:
      - description: Rename Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.SettingRenameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.SettingDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Rename setting
      tags:
      - Setting
  /api/settings:
    get:
      description: Get setting list for current user with pagination and keyword filtering
      parameters:
      - description: Keyword // 关键词
        example: User/
        in: query
        name: keyword
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      - description: Page number // 页码
        in: query
        name: page
        type: integer
      - description: Page size // 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/app.ListRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SettingDTO'
                        type: array
                    type: object
              type: object
      security:
      - UserAuthToken: []
      summary: Get setting list
      tags:
      - Setting
...
securityDefinitions:
  ShareAuthToken:
    in: header
    name: Share-Token
    type: apiKey
  UserAuthToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"